Remove Signon Plek Override In AWS
This override was added during the migration from AWS to Carrenza, to ensure
that apps hosted by different providers would still be able to communicate with
Signon.

Now that all apps are in AWS, we want them to use the internal load balancer to
connect to Signon. To do this, we must ensure they have
signon.production.govuk-internal.digital set as their Signon URL.

This is an incident action item, you can find more information about this here:
https://docs.google.com/document/d/1npjn9yQ0z_p9ie2RBikt-VpigJWoAz76pgHwNOMlZKM/edit#